iceberg/requirements.txt
pyiceberg
pyiceberg[s3fs,pyarrow]
boto3
python-dotenv
requests
